[common] Optimize LogRecordBatchStatistics.getNullCounts() to return int[] instead of Long[] (#3054)

+++ fluss-common/src/main/java/org/apache/fluss/predicate/And.java
@@ -51,7 +51,7 @@ public class And extends CompoundPredicate.Function {
             long rowCount,
             InternalRow minValues,
             InternalRow maxValues,
-            Long[] nullCounts,
+            int[] nullCounts,
             List<Predicate> children) {
         for (Predicate child : children) {
             if (!child.test(rowCount, minValues, maxValues, nullCounts)) {
