+++ fluss-common/src/main/java/org/apache/fluss/predicate/Predicate.java
@@ -45,7 +45,7 @@ public interface Predicate extends Serializable {
      * @return return true is likely to hit (there may also be false positives), return false is
      *     absolutely not possible to hit.
      */
-    boolean test(long rowCount, InternalRow minValues, InternalRow maxValues, Long[] nullCounts);
+    boolean test(long rowCount, InternalRow minValues, InternalRow maxValues, int[] nullCounts);
 
     /**
      * @return the negation predicate of this predicate if possible.
